requirements.txt
Django==5.0
pillow
requests
django-cron
djangorestframework
mysqlclient
djangorestframework-simplejwt